backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
pytesseract==0.3.10
spacy==3.7.2
nltk==3.8.1
chromadb==0.4.15
pymongo==4.6.0
firebase-admin==6.2.0
boto3==1.34.0
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
pillow==10.1.0
numpy==1.24.3
scikit-learn==1.3.2